backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9

langchain==0.2.1
langchain-community==0.2.1
langchain-google-genai==1.0.5
langgraph==0.0.69
google-generativeai==0.5.4

pypdf==4.2.0
pdfplumber==0.11.0
python-docx==1.1.2

chromadb==0.5.0
rank-bm25==0.2.2
sentence-transformers==3.0.0

numpy==1.26.4
pandas==2.2.2
tiktoken==0.7.0
tenacity==8.3.0
httpx==0.27.0
aiofiles==23.2.1

structlog==24.1.0
prometheus-client==0.20.0

pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0